third_party/blink/web_tests/external/wpt/editing/other/formatblock-preserving-selection.tentative_styleWithCSS=true_block=blockquote-expected.txt

This is a testharness.js-based test.
[FAIL] Preserve selection after formatBlock with blockquote at <table><tr><td>a[bc</td></tr></table><div>de]f</div>
  assert_equals: Result: <blockquote><table><tbody><tr><td><blockquote>abc</blockquote></td></tr></tbody></table>def</blockquote> expected "bcde" but got "bcdef"
Harness: the test ran to completion.
